[core] Add i18n reducer to core, remove i18n package
