--- packages/i18n/README.md
@@ -1,26 +1,0 @@
-# JSONForms - More Forms. Less Code
-### Complex Forms in the blink of an eye
-
-JSONForms eliminates the tedious task of writing fully-featured forms by hand by leveraging the capabilities of JSON, JSON Schema and Javascript.
-
-# I18n Package
-This is the JSONForms i18n package. This package provides reducers and actions to support i18n.
-
-You can combine this package with any renderer you want, for example the [Material Renderers](https://www.npmjs.com/package/@jsonforms/material-renderers) or the [Vanilla Renderers](https://www.npmjs.com/package/@jsonforms/vanilla-renderers).
-
-See the [Example Package](https://www.npmjs.com/package/@jsonforms/examples) on how to integrate JSONForms with your application.
-
-Check https://www.npmjs.com/search?q=%40jsonforms for all published JSONForms packages.
-
-# License
-The JSONForms project is licensed under the MIT License. See the [LICENSE file](https://github.com/eclipsesource/jsonforms/blob/master/LICENSE) for more information.
-
-# Roadmap
-Our current roadmap is available [here](https://github.com/eclipsesource/jsonforms/blob/master/ROADMAP.md).
-
-# Development
-JSONForms is developed by [EclipseSource](https://eclipsesource.com).
-We are always very happy to have contributions, whether for trivial cleanups or big new features.
-
-# Migration
-If you are already using JSONForms 1, check our [migration guide](https://github.com/eclipsesource/jsonforms/blob/master/MIGRATION.md).
