--- packages/i18n/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@jsonforms/i18n",
-  "version": "2.0.12-rc.4",
-  "description": "i18n module of JSON Forms",
-  "repository": "https://github.com/eclipsesource/jsonforms",
-  "bugs": "https://github.com/eclipsesource/jsonforms/issues",
-  "homepage": "http://jsonforms.io/",
-  "private": true,
-  "license": "MIT",
-  "directories": {
-    "src": "src",
-    "test": "test"
-  },
-  "main": "./lib/index.js",
-  "typings": "./lib/index.d.ts",
-  "scripts": {
-    "bundle": "../../node_modules/.bin/webpack --config ./webpack.build.js --env=production --display-error-details",
-    "build": "../../node_modules/.bin/tsc",
-    "clean": "rm -rf lib coverage dist .nyc_output 2> /dev/null",
-    "lint": "../../node_modules/.bin/tslint --project ../../tslint.json --exclude src/models/jsonSchema.ts",
-    "report": "../../node_modules/.bin/nyc report --reporter=html",
-    "test": "../../node_modules/.bin/tsc --project test/tsconfig.test.json && ../../node_modules/.bin/ava",
-    "test-cov": "../../node_modules/.bin/tsc --project test/tsconfig.test.json && ../../node_modules/.bin/nyc ../../node_modules/.bin/ava",
-    "doc": "../../node_modules/.bin/typedoc --name 'JSON Forms I18n' --mode file --excludeExternals --theme ../../typedoc-jsonforms --out docs src"
-  },
-  "ava": {
-    "verbose": true,
-    "require": [
-      "babel-register"
-    ],
-    "babel": "inherit"
-  },
-  "nyc": {
-    "reporter": [
-      "lcov",
-      "text-summary"
-    ]
-  },
-  "dependencies": {
-    "@jsonforms/core": "^2.0.12-rc.4",
-    "es6-shim": "^0.35.3",
-    "json-refs": "^3.0.10",
-    "moment": "^2.20.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-redux": "^5.0.6",
-    "redux": "^3.7.2"
-  },
-  "devDependencies": {
-    "@jsonforms/test": "^2.0.12-rc.4",
-    "@jsonforms/vanilla-renderers": "^2.0.12-rc.4",
-    "jsdom": "^11.9.0",
-    "jsdom-global": "^3.0.2"
-  }
-}
