--- packages/i18n/test/tsconfig.test.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "../dist/test",
-    "target": "es6",
-    "inlineSourceMap": true
-  },
-  "include": [
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude":[
-    "node_modules"
-  ],
-  "files": [
-    "../src/index.ts"
-  ]
-}
