--- packages/i18n/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.base",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "sourceMap": true
-  },
-  "exclude":[
-    "node_modules"
-  ],
-  "files": [
-    "./src/index.ts"
-  ]
-}
