--- packages/i18n/webpack.build.js
@@ -1,9 +1,0 @@
-const merge = require('webpack-merge');
-const baseConfig = require('../../webpack/webpack.build.base.js');
-
-module.exports = merge(baseConfig, {
-    output: {
-        filename: "jsonforms-i18n.js",
-        library: "JSONFormsI18n"
-      },
-});
